Small refactor to bloomsky
