+++ homeassistant/components/bloomsky/binary_sensor.py
@@ -15,7 +15,7 @@ import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import DOMAIN
+from .const import DATA
 
 SENSOR_TYPES = {"Rain": BinarySensorDeviceClass.MOISTURE, "Night": None}
 
@@ -40,7 +40,7 @@ def setup_platform(
         return
 
     sensors = config[CONF_MONITORED_CONDITIONS]
-    bloomsky = hass.data[DOMAIN]
+    bloomsky = hass.data[DATA]
 
     for device in bloomsky.devices.values():
         for variable in sensors:
